# packages/api-types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Dataplat API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/chunks/query": {
      "post": {
        "tags": [
          "chunks"
        ],
        "summary": "Query Chunks",
        "description": "Execute a DataFusion SQL filter on the Lance chunks table.\n\nAuth required (F-008).\n\nIMPORTANT \u2014 no per-user row scoping:\n  The handler requires a valid bearer token but does NOT inject owner-scoping\n  into the Lance filter.  Callers are responsible for scoping via the filter\n  field (e.g. \"source_id = 42\").  Repository-level ACL on Lance is deferred\n  to post-MVP (design doc \u00a711.6).\n\nReturns:\n  items \u2014 up to `limit` matching chunk rows (all 24 fields; unselected\n          fields are None when `columns` is specified).\n  total \u2014 count of ALL rows matching the filter (ignores limit/offset).",
        "operationId": "query_chunks_api_chunks_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChunkQueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChunkQueryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/chunks/aggregate": {
      "post": {
        "tags": [
          "chunks"
        ],
        "summary": "Aggregate Chunks",
        "description": "Compute grouped statistics over the Lance chunks table.\n\nAuth required (F-008).  No per-user row scoping (\u00a711.6 deferred).\n\nAll matching rows (subject to filter) are loaded into process memory for\ngrouping \u2014 callers should apply a filter to avoid full-table scans on\nlarge datasets.  Post-MVP: push GROUP BY to DuckDB/DataFusion SQL.",
        "operationId": "aggregate_chunks_api_chunks_aggregate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChunkAggregateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChunkAggregateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/chunks/distribution": {
      "post": {
        "tags": [
          "chunks"
        ],
        "summary": "Distribution Chunks",
        "description": "Compute a value distribution histogram for one column.\n\nAuth required (F-008).  No per-user row scoping (\u00a711.6 deferred).\n\nColumn type is auto-detected from the PyArrow schema:\n  - Floating-point / integer \u2192 numeric equal-width histogram (bins buckets).\n  - String (utf8/large_utf8) \u2192 categorical value counts, count descending.\n  - Other types (bool, list, timestamp, \u2026) \u2192 HTTP 400.\n\nAll matching rows for the target column are loaded into process memory \u2014\ncallers should apply a filter to avoid full-table scans on large datasets.\nPost-MVP: push aggregation to DuckDB/DataFusion SQL.",
        "operationId": "distribution_chunks_api_chunks_distribution_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChunkDistributionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChunkDistributionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/chunks/{chunk_id}/lineage": {
      "get": {
        "tags": [
          "chunks"
        ],
        "summary": "Get Chunk Lineage",
        "description": "Walk the augmented_from chain for a chunk and return source + canonical document_variant.\n\nAuth required (F-008).  No per-user row scoping (\u00a711.6 deferred).\n\nSteps:\n  1. Fetch the requested chunk from Lance (asyncio.to_thread).\n  2. Walk the augmented_from chain iteratively (tip \u2192 root):\n       - Depth cap: _MAX_LINEAGE_DEPTH (32) via for/else.\n       - Cycle guard: seen_ids set detects re-entry before HTTP 500.\n       - Root check (augmented_from is None) fires BEFORE the depth cap\n         so a chain of exactly 32 valid entries still returns HTTP 200.\n  3. Fetch the source record from Postgres (async SQLAlchemy select).\n  4. Fetch the canonical document_variant for that source (async, nullable).\n  5. Return ChunkLineageResponse.\n\nError responses:\n  404 \u2014 chunk_id not found in Lance, or source_id not in Postgres.\n  400 \u2014 Lance/DataFusion query error.\n  500 \u2014 cycle detected, depth cap exceeded, broken augmented_from reference,\n         or null source_id on root chunk (data integrity violations).",
        "operationId": "get_chunk_lineage_api_chunks__chunk_id__lineage_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "chunk_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 256,
              "title": "Chunk Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChunkLineageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/chunks/{chunk_id}": {
      "get": {
        "tags": [
          "chunks"
        ],
        "summary": "Get Chunk By Id",
        "description": "Fetch a single chunk by its chunk_id.\n\nAuth required (F-008).  No per-user row scoping (\u00a711.6 deferred).\n\nReturns HTTP 404 if no chunk with the given chunk_id exists in the Lance\ntable.  Returns HTTP 400 if Lance/DataFusion raises any error.",
        "operationId": "get_chunk_by_id_api_chunks__chunk_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "chunk_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 256,
              "title": "Chunk Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChunkRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChunkQueryRequest": {
        "properties": {
          "filter": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Filter"
          },
          "columns": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Columns"
          },
          "limit": {
            "type": "integer",
            "maximum": 1000.0,
            "minimum": 1.0,
            "title": "Limit",
            "default": 100
          },
          "offset": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Offset",
            "default": 0
          }
        },
        "type": "object",
        "title": "ChunkQueryRequest",
        "description": "Request body for POST /api/chunks/query.\n\nfilter   \u2014 DataFusion SQL predicate fragment applied to the Lance chunks\n           table (e.g. \"source_id = 42\", \"attr_quality_score > 0.8\").\n           None / omitted means no filter (return all rows, subject to\n           limit/offset).  Max 1000 chars.\ncolumns  \u2014 Optional list of column names to project.  None = all 24 columns.\n           Unknown column names cause a 400 (DataFusion parse error).\nlimit    \u2014 Max rows per page (1\u20131000, default 100).\noffset   \u2014 Row offset for pagination (default 0)."
      },
      "ChunkQueryResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ChunkRead"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "ChunkQueryResponse",
        "description": "Paginated response for POST /api/chunks/query."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/token"
          }
        }
      }
    }
  }
}